% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTools.R
\name{fdp}
\alias{fdp}
\title{Title
FDP Calculate}
\usage{
fdp(selected, H0)
}
\arguments{
\item{selected}{a vector indicating the indexes of the selected hypotheses.}

\item{H0}{a vector indicating the underlying states of the hypotheses.}
}
\value{
a numerical value representing the false discovery proportion.
}
\description{
Calculate the false discovery proportion.
}
