% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMirror.R
\docType{package}
\name{JMirror}
\alias{JMirror}
\title{JMirror: A Package for Joint Mirror Procedure}
\description{
The JMirror package provides two important functions for implementing joint mirror procedure: JointMirror.R and JointMirror.Qvalue.
}
