% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointMirrorR.R
\name{JointMirror.Qvalue}
\alias{JointMirror.Qvalue}
\title{Title
Closed FDP estimate}
\usage{
JointMirror.Qvalue(Pval, JMirrorS)
}
\arguments{
\item{Pval}{a m*K matrix for p-values;}

\item{JMirrorS}{is a module storing the results of joint mirror procedure; usually, this is the result of JMirrorR() with trgt.fdr.level=0.}
}
\value{
list
Unmask.Order is the unmasking order of each hypothesis.
is.RejSide represents whether the hypothesis is in the rejection side.
FDP.est.Total represents the adjusted FDP estimates that is decreasing with respect to the unmasking order.
}
\description{
If there are several target fdr levels to compare, we can set trgt.fdr.level=0 in JointMirror.R() and input result here. We can directly compare the FDP estimates and target fdr level to obtain discoveries.
}
\examples{
set.seed(10)
m <- 1000
mu1H10 <- 1.5
mu1H11 <- 2
mu2H01 <- 2.5
mu2H11 <- 3
pi.seq <- c(0.4,0.2,0.2,0.2)
H <- rmultinom(m,1,pi.seq)
H0 <- colSums(H[1:3,])
mu1 <- mu2 <- rep(0,m)

mu1[H[2,]==1] <- mu1H10
mu1[H[4,]==1] <- mu1H11
mu2[H[3,]==1] <- mu2H01
mu2[H[4,]==1] <- mu2H11

X1 <- rnorm(m,mu1)
X2 <- rnorm(m,mu2)

p1 <- 2*(pnorm(-abs(X1)))
p2 <- 2*(pnorm(-abs(X2)))

Pval = cbind(p1,p2)
JM.Product.Res <- JointMirror.R(Pval,rank.Mode = "Product",trgt.fdr.level=0)
JM.Product.Qval <- JointMirror.Qvalue(Pval,JM.Product.Res$JMirrorS)
trgt.fdr.level <- 0.2
FDP.est <- JM.Product.Qval$FDP.est.Total
is.RejSide <- JM.Product.Qval$is.RejSide
JM.Product.sel <- which(FDP.est<=trgt.fdr.level & is.RejSide)
c(fdp(JM.Product.sel,H0),Pow(JM.Product.sel,H0))
}
