% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTools.R
\name{Pow}
\alias{Pow}
\title{Title
Power Calculate}
\usage{
Pow(selected, H0)
}
\arguments{
\item{selected}{a vector indicating the indexes of the selected hypotheses.}

\item{H0}{a vector indicating the underlying states of the hypotheses.}
}
\value{
a numerical value representing the true positive percentage.
}
\description{
Calculate the true positive percentage.
}
