% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp Object for conducting Joint Mirror Procedure.}
\alias{Rcpp Object for conducting Joint Mirror Procedure.}
\title{Rcpp Object for conducting Joint Mirror Procedure.}
\description{
Rcpp Object for conducting Joint Mirror Procedure.
}
\section{Fields}{

\describe{
\item{\code{mult}}{Multiply by another Double object \itemize{
\item Parameter: Pval_ is the p-value matrix, Hker.inv.chol is the bandwidth matrix.
}}
}}

